+++ src/components/index.ts
@@ -19,6 +19,7 @@ export { default as MdIcons } from './utils/icons/icons-mdi';
 export { Animation } from './utils/getAnimationStyles';
 export { takeIf } from './utils/helpers';
 export { generateId } from './utils/generateId';
+export { useHotkeys } from './utils/useHotkeys';
 export { Accordion } from './Accordion/Accordion';
 export type { AccordionProps } from './Accordion/Accordion';
 export { AccordionItem } from './Accordion/AccordionItem/AccordionItem';